PluginMeta and co moved to pyworkflow.plugin Catch first natural xmipp binding import and inform if xmipp not installed Cancel xmipp installation at "scipion" script.
